Add AUTOBOOT option.
AUTOBOOT is a boot.cfg option that allows
one to boot an arbitrary section (boot option)
of a boot.cfg file automatically,
without showing the kexecboot menu.

The AUTOBOOT option is useful when a user
wants kexecboot to default on a certain
boot option and only show the boot menu when
the user explicitly requests it.

When AUTOBOOT option is used and the user
wishes kexecboot to show the boot menu,
the user should simply hold any button
or touchscreen during boot and kexecboot
will ignore the AUTOBOOT option and show
the boot menu.

To use the AUTOBOOT option, simply add
"AUTOBOOT" keyword to a boot section
you want kexecboot to default on.

If more than one boot section has the
AUTOBOOT keyword, the first one found
during device scan will be used.